Add files via upload
changes creates a "open" button when pressed, the pdf is downloaded to user's computer. 
I had tried having the PDF open in the UI --> was able to create a pop up box but the content of the PDF would not show
I have also tried having the PDF pop up in a new tab --> faced security errors that had the content blocked.

Please let me know if you have any thoughts! Thank you